Updated the README
Updated the README to include some information about the program and the
source code. With notes for people who want to run it on modern versions
of Windows, such as Vista/7/8+

Also in the readme are links to the posts on both Twitter and Mastadon
where I showed I did this.

Signed-off-by: Rebecca White <rwhitedev4@gmail.com>